Revert if conditions

+++ lumina_next_t2i_mini/train_vae_swap_layers.py
@@ -251,7 +251,7 @@ def main(args):
         model_kwargs = dict(cap_feats=cap_feats.contiguous(), cap_mask=cap_mask.contiguous())
 
         # Forward pass
-        with torch.cuda.amp.autocast():
+        with torch.cuda.amp.autocast(dtype=torch.float32):
             loss_dict = training_losses(model, latent, latent, model_kwargs)
         opt.zero_grad()
         loss = loss_dict["loss"].sum()
